Brush Free()
Brush insert into the new tri tree stuff

map dump in BrushList now works better with Quark (sees it as a GrogLibs map)

GBSPSide free()

TriTree DebugDrawChoice

Added QuadTree Block stuff to TriTree, didn't really help

BasicModelHelper now operating from float seconds

Same for LightHelper and StaticHelper

TriggerHelper using float millisecond deltas, need to look at that

Zone Model stuff using float deltas, trigger time since triggered using a double now in case the time goes long

+++ BSPZone/Helpers/StaticHelper.cs
@@ -164,7 +164,7 @@ namespace BSPZone
 		}
 
 
-		public void Update(int msDelta)
+		public void Update(float msDelta)
 		{
 			Debug.Assert(msDelta > 0);
